requirements.txt
CTkMessagebox==2.5
customtkinter==5.2.2
numpy==1.26.4
Pillow==10.2.0